gdb/frame: Refine stack switch heuristic
GDB includes a heuristic that terminates unwinding if GDB comes across
a stack frame transition that looks like it's going in the wrong
direction. Of course, in an application with perfect unwind info, such
a heuristic should not be required, but alas in such a world we live not.

Quite inconveniently, there are libraries that do cause the stack
to look as if its going in the wrong direction, for legitimate reasons,
e.g. because they're implementing segmented stacks or because they're
doing some sort of instrumentation that should run on a separate stack.
The trouble with this heuristic is that even if these libraries declare
their CFI carefully and perfectly, gdb will still terminate unwinding.

There is currently one workaround to disable this heuristic: Because
stack switching is used in the gcc segmented stack implementation,
the heuristic inside gdb is disabled if the function happens to be
called __morestack (which is the name of the gcc function
implementing the stack segmentation). Because of this, a host of
applications have started calling their functions __morestack in
an attempt to get gdb to disable this heuristic (e.g. [1][2]).

I was about to add this workaround to another application, but I
figured it might be worth seeing if we can do better in gdb instead.
The workaround works, but having a bunch of unrelated non-descript
__morestack functions in the backtrace, rather than a more useful
indication of why the stack is switching is not a great debugging
experience. It also complicates debugging of applications that have
CFI information available (because it was in memory when the
application crashed), but no symbols.

Here I propose the following extension to GDB's heuristic: If the
CFA of a frame is declared using DW_CFA_def_cfa_expression, i.e.
is a DWARF expression, then gdb disables the heuristic, because
clearly something fancy is going on with the stacks. In most cases
complicated stack switchers will likely be using this form of
the CFI since the computation of the CFA is non-trivial after
the stack has been switched to the new location (likely involving
reading a frame on the new stack to determine the old rsp). However,
even if the CFA computation is not as complicated (e.g. because
the stack switcher happens to save the CFA in a register), this
gives applications an easy low-overhead workaround to mark their
functions as a stack switcher without having to rename their symbols.

I think this is a reasonable tradeoff. DW_CFA_def_cfa_expression is
reasonably rare outside of code that does messy things with stack
computation, so it doesn't defang the heuristic while giving a clear
rule for when the heuristic is disabled. An alternatively could
be a more explicit way to mark stack switchers in the CFI that
does not involve requiring a symbol lookup, but that would probably
require more coordination with the standards authors and CFI
producers, which doesn't seem worthwhile at this juncture.

[1] https://github.com/rr-debugger/rr/blob/1a774d0b5946a19b291756092be699d5295c0dc1/src/preload/syscall_hook.S#L370-L382
[2] https://github.com/edef1c/libfringe/blob/master/src/arch/x86_64.rs#L64-L68

Signed-off-by: Keno Fischer <keno@juliacomputing.com>